MAR-60 create route group for docs layout
